[AOT Refactor] logging utils
[ghstack-poisoned]
